+++ src/amd/utils.py
@@ -8,8 +8,13 @@ import numba
 from scipy.spatial.distance import squareform
 
 __all__ = [
-    'diameter', 'cellpar_to_cell', 'cellpar_to_cell_2D', 'cell_to_cellpar',
-    'cell_to_cellpar_2D', 'neighbours_from_distance_matrix', 'random_cell'
+    'diameter',
+    'cellpar_to_cell',
+    'cellpar_to_cell_2D',
+    'cell_to_cellpar',
+    'cell_to_cellpar_2D',
+    'neighbours_from_distance_matrix',
+    'random_cell'
 ]